Fix investigator name display in complaints and merge status update into dashboard
- Fixed 'Unknown Investigator' issue in complaints table by using investigator_name/investigator_email from backend response
- Updated Complaint model to_dict() to use full_name or email fallback
- Merged Status Update functionality into investigator dashboard (removed separate section)
- Fixed dashboard charts data validation and display
- Improved error handling and data validation in dashboard components
- Added InvestigatorMap component for admin panel
- Enhanced synthetic fraud data generation to involve multiple accounts
- Fixed evidence upload authentication and superadmin notifications
- Improved role-based access control for AI analysis reports

+++ backend/app/api/v1/schemas.py
@@ -177,8 +177,8 @@ class IncidentReportRequest(BaseModel):
 
 
 class TransactionSummary(BaseModel):
-    total_in: int
-    total_out: int
+    total_in: float
+    total_out: float
     tx_count: int
     unique_senders: int
     unique_receivers: int
@@ -188,7 +188,7 @@ class TransactionSummary(BaseModel):
 class GraphEdge(BaseModel):
     from_address: str = Field(alias="from")
     to_address: str = Field(alias="to")
-    amount: int
+    amount: float
     
     class Config:
         populate_by_name = True
@@ -196,7 +196,7 @@ class GraphEdge(BaseModel):
 
 class TimelineEntry(BaseModel):
     time: str
-    amount: int
+    amount: float
     timestamp: str
